webkit/LayoutTests/fast/replaced/replaced-percent-height-expected.txt



This test checks that replaced elements with percentage heights computes to 'auto' in standard mode, if the height of the containing block is not specified explicitly.
See bug #37169.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getWidth('image1') == getHeight('image1') is true
PASS getHeight('image2') == '100px' is true
PASS successfullyParsed is true

TEST COMPLETE

